requirements.txt
youtube-search-python==1.6.6
youtube-transcript-api==1.0.3
httpx==0.28.1
pydantic==2.11.1
python-dotenv==1.1.0
numpy==2.2.4
rich==14.0.0
fastapi==0.135.3
uvicorn==0.44.0
